// src/types/timesheet/model/timesheet-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zero consult timesheet - OpenAPI 3.2
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const TimesheetStatus = {
    InProgress: 'In progress',
    Approved: 'Approved',
    Rejected: 'Rejected',
} as const;

export type TimesheetStatus = typeof TimesheetStatus[keyof typeof TimesheetStatus];